refactor(grpc): introduce injectable GrpcTransport, extract native transport
